@@ -1,10 +1,6 @@
 import adapter from '@sveltejs/adapter-node';
 
 /** @type {import('@sveltejs/kit').Config} */
-const config = {
-	kit: {
-		adapter: adapter()
-	}
-};
+const config = { kit: { adapter: adapter() } };
 
 export default config;
